Feature Projeto

+++ src/app/album-table-item/album-table-item.component.ts
@@ -3,7 +3,7 @@ import { Album } from '../models/album';
 import { AlbumService } from '../services/album.service';
 
 @Component({
-  selector: 'tbody',
+  selector: 'app-album-table-item',
   templateUrl: './album-table-item.component.html',
   styleUrls: ['./album-table-item.component.css']
 })
@@ -18,12 +18,16 @@ export class AlbumTableItemComponent implements OnInit {
   }
 
   remove(album: Album) {
-    this.albumService.delete(album).subscribe({
-      next: () => this.albumService.getAll().subscribe({
-        next: (albums: Album[]) => this.albumList = albums,
+    let answer = confirm(`Deseja remover o álbum '${album.title}'?`);
+
+    if(answer) {
+      this.albumService.delete(album).subscribe({
+        next: () => this.albumService.getAll().subscribe({
+          next: (albums: Album[]) => this.albumList = albums,
+          error: (error) => alert(error.message)
+        }),
         error: (error) => alert(error.message)
-      }),
-      error: (error) => alert(error.message)
-    });
+      });
+    }
   }
 }
